+++ lotti/lib/sync/inbox_service.dart
@@ -301,7 +301,8 @@ class SyncInboxService {
         _observingClient = MailClient(account, isLogEnabled: false);
 
         await _observingClient!.connect();
-        await _observingClient!.selectInbox();
+        await _observingClient!
+            .selectMailboxByPath(imapConfig.folder ?? 'INBOX');
 
         _observingClient!.eventBus
             .on<MailLoadEvent>()